Add GitHub Actions workflow for CI/CD
The 'main.yml' action workflow was added to the '.github/workflows' directory to automate the CI/CD process. This workflow triggers on every push to the main branch, runs on the latest version of Ubuntu, sets up Node.js 16.x and installs pnpm, dependencies, and builds the project finally deploying to GitHub Pages. This addition ensures that the development process is more streamlined and less prone to manual error.