Add loading representation to User Details component
Introduced `isLoading` field to the User Details component for handling its loading state. This new field is announced and managed in the global state to show a loading message before user data loaded completely. As a result, we can provide a response to the user during the long-loading processes and improve user experience.